feat: added fixes to comply for front end

+++ backend/src/main/java/com/example/ticketingsystem/dto/response/AgentPerformanceResponse.java
@@ -13,5 +13,5 @@ public class AgentPerformanceResponse {
     private long resolvedTickets;
     private long closedTickets;
     private Double averageResolutionTimeHours;
-    private Integer workload;
+    private long workload;
 }
